flows/requirements.txt
pandas
pyarrow
confluent-kafka>=1.9.0
scikit-learn
lightgbm
mlflow>=2.12.2
python-dotenv
apache-airflow>=2.8.1
psycopg2-binary
duckdb
